Conflicts:
	newscoop/library/Newscoop/Entity/Article.php

+++ newscoop/admin-files/plugins/manage.php
@@ -21,6 +21,10 @@ if (Input::Get('save')) {
     // delete from DB those which were uninstalled
     foreach (CampPlugin::getAll() as $CampPlugin) {
         if (!array_key_exists($CampPlugin->getName(), $p_plugins)) {
+            if (strpos($CampPlugin->getName(), '/') !== false) {
+                continue;
+            }
+            
             $CampPlugin->delete();
         }
     }
